rename base component
