+++ frontend/components/global/BaseExpansionPanels.vue
