+++ frontend/types/components.d.ts
@@ -12,8 +12,8 @@ import type BaseButton from "@/components/global/BaseButton.vue";
 import type BaseButtonGroup from "@/components/global/BaseButtonGroup.vue";
 import type BaseCardSectionTitle from "@/components/global/BaseCardSectionTitle.vue";
 import type BaseDialog from "@/components/global/BaseDialog.vue";
-import type BaseDisclosure from "@/components/global/BaseDisclosure.vue";
 import type BaseDivider from "@/components/global/BaseDivider.vue";
+import type BaseExpansionPanels from "@/components/global/BaseExpansionPanels.vue";
 import type BaseOverflowButton from "@/components/global/BaseOverflowButton.vue";
 import type BasePageTitle from "@/components/global/BasePageTitle.vue";
 import type BaseStatCard from "@/components/global/BaseStatCard.vue";
@@ -54,8 +54,8 @@ declare module "vue" {
     BaseButtonGroup: typeof BaseButtonGroup;
     BaseCardSectionTitle: typeof BaseCardSectionTitle;
     BaseDialog: typeof BaseDialog;
-    BaseDisclosure: typeof BaseDisclosure;
     BaseDivider: typeof BaseDivider;
+    BaseExpansionPanels: typeof BaseExpansionPanels;
     BaseOverflowButton: typeof BaseOverflowButton;
     BasePageTitle: typeof BasePageTitle;
     BaseStatCard: typeof BaseStatCard;
